+++ test/OpenTelemetry.Instrumentation.ConfluentKafka.Tests/EventSourceTests.cs
@@ -1,0 +1,13 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+using OpenTelemetry.Tests;
+
+namespace OpenTelemetry.Instrumentation.ConfluentKafka.Tests;
+
+public class EventSourceTests
+{
+    [Fact]
+    public void EventSourceTests_ConfluentKafkaInstrumentationEventSource()
+        => EventSourceTestHelper.ValidateEventSourceIds<ConfluentKafkaInstrumentationEventSource>();
+}