+++ src/OpenTelemetry.Instrumentation.ConfluentKafka/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added `messaging.kafka.cluster.id` span attribute.
+  ([#4739](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4739))
+
 ## 0.2.0-alpha.2
 
 Released 2026-Jul-17